+++ tests/handlers/allowSpendsHandler.test.ts
@@ -7,191 +7,17 @@ import {
 } from "../testUtils";
 import {
   data_addresses,
+  data_dag_allow_spends,
+  data_dag_expired_spend_transactions,
+  data_dag_spend_transactions,
   data_global_snapshots,
+  data_metagraph_allow_spends,
+  data_metagraph_expired_spend_transactions,
   data_metagraph_snapshots,
+  data_metagraph_spend_transactions,
   data_metagraphs,
   prisma,
 } from "../../prisma/seed";
-
-export const data_dag_allow_spends = [
-  {
-    hash: "allowSpendHash1",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[1].address,
-    amount: 1000n,
-    fee: 5n,
-    last_valid_epoch_progress: 100n,
-    ordinal: 1n,
-    snapshot_hash: data_global_snapshots[0].hash,
-    round_id: "11111111-1111-1111-1111-111111111111",
-    created_at: new Date("2024-01-01T10:00:00Z"),
-    updated_at: new Date("2024-01-01T10:00:00Z"),
-  },
-  {
-    hash: "allowSpendHash2",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[2].address,
-    amount: 2000n,
-    fee: 10n,
-    last_valid_epoch_progress: 600n,
-    ordinal: 2n,
-    snapshot_hash: data_global_snapshots[0].hash,
-    round_id: "22222222-2222-2222-2222-222222222222",
-    created_at: new Date("2024-01-02T10:00:00Z"),
-    updated_at: new Date("2024-01-02T10:00:00Z"),
-  },
-  {
-    hash: "allowSpendHash3",
-    source_addr: data_addresses[2].address,
-    destination_addr: data_addresses[1].address,
-    amount: 300n,
-    fee: 15n,
-    last_valid_epoch_progress: 700n,
-    ordinal: 3n,
-    snapshot_hash: data_global_snapshots[1].hash,
-    round_id: "33333333-3333-3333-3333-333333333333",
-    created_at: new Date("2024-01-03T10:00:00Z"),
-    updated_at: new Date("2024-01-03T10:00:00Z"),
-  },
-  {
-    hash: "allowSpendHash4",
-    source_addr: data_addresses[2].address,
-    destination_addr: data_addresses[3].address,
-    amount: 300n,
-    fee: 15n,
-    last_valid_epoch_progress: 700n,
-    ordinal: 4n,
-    snapshot_hash: data_global_snapshots[0].hash,
-    round_id: "33333333-3333-3333-3333-333333333333",
-    created_at: new Date("2024-01-03T10:00:00Z"),
-    updated_at: new Date("2024-01-03T10:00:00Z"),
-  },
-];
-
-export const data_dag_spend_transactions = [
-  {
-    hash: "spendTxHash1",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[1].address,
-    amount: 1000n,
-    allow_spend_ref: data_dag_allow_spends[0].hash,
-    snapshot_hash: data_global_snapshots[2].hash,
-    created_at: new Date("2024-01-05T10:00:00Z"),
-    updated_at: new Date("2024-01-05T10:00:00Z"),
-  },
-];
-
-export const data_dag_expired_spend_transactions = [
-  {
-    hash: "expiredSpendTxHash1",
-    source_addr: data_addresses[2].address,
-    amount: 3000n,
-    allow_spend_ref: data_dag_allow_spends[1].hash,
-    snapshot_hash: data_global_snapshots[2].hash,
-    created_at: new Date("2024-01-10T10:00:00Z"),
-    updated_at: new Date("2024-01-10T10:00:00Z"),
-  },
-];
-
-export const data_metagraph_allow_spends = [
-  {
-    metagraph_id: data_metagraphs[0].id,
-    hash: "metaAllowSpendHash1",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[1].address,
-    amount: 500n,
-    fee: 3n,
-    last_valid_epoch_progress: 50n,
-    ordinal: 1n,
-    snapshot_hash: data_metagraph_snapshots[0].hash,
-    round_id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
-    created_at: new Date("2024-01-01T11:00:00Z"),
-    updated_at: new Date("2024-01-01T11:00:00Z"),
-  },
-  {
-    metagraph_id: data_metagraphs[0].id,
-    hash: "metaAllowSpendHash2",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[1].address,
-    amount: 500n,
-    fee: 3n,
-    last_valid_epoch_progress: 50n,
-    ordinal: 2n,
-    snapshot_hash: data_metagraph_snapshots[1].hash,
-    round_id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
-    created_at: new Date("2024-01-01T11:00:00Z"),
-    updated_at: new Date("2024-01-01T11:00:00Z"),
-  },
-  {
-    metagraph_id: data_metagraphs[0].id,
-    hash: "metaAllowSpendHash3",
-    source_addr: data_addresses[2].address,
-    destination_addr: data_addresses[3].address,
-    amount: 750n,
-    fee: 4n,
-    last_valid_epoch_progress: 60n,
-    ordinal: 3n,
-    snapshot_hash: data_metagraph_snapshots[1].hash,
-    round_id: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb",
-    created_at: new Date("2024-01-02T11:00:00Z"),
-    updated_at: new Date("2024-01-02T11:00:00Z"),
-  },
-];
-
-export const data_metagraph_spend_transactions = [
-  {
-    metagraph_id: data_metagraphs[0].id,
-    hash: "metaSpendTxHash1",
-    source_addr: data_addresses[0].address,
-    destination_addr: data_addresses[1].address,
-    amount: 500n,
-    allow_spend_ref: data_metagraph_allow_spends[0].hash,
-    snapshot_hash: data_metagraph_snapshots[0].hash,
-    created_at: new Date("2024-01-03T11:00:00Z"),
-    updated_at: new Date("2024-01-03T11:00:00Z"),
-  },
-];
-
-export const data_metagraph_expired_spend_transactions = [
-  {
-    metagraph_id: data_metagraphs[0].id,
-    hash: "metaExpiredTxHash1",
-    source_addr: data_addresses[2].address,
-    amount: 750n,
-    allow_spend_ref: data_metagraph_allow_spends[2].hash,
-    snapshot_hash: data_metagraph_snapshots[1].hash,
-    created_at: new Date("2024-01-04T11:00:00Z"),
-    updated_at: new Date("2024-01-04T11:00:00Z"),
-  },
-];
-
-const seedData = async () => {
-  await prisma.dag_allow_spends.createMany({
-    data: data_dag_allow_spends,
-  });
-
-  await prisma.dag_spend_transactions.createMany({
-    data: data_dag_spend_transactions,
-  });
-
-  await prisma.dag_expired_spend_transactions.createMany({
-    data: data_dag_expired_spend_transactions,
-  });
-
-  await prisma.metagraph_allow_spends.createMany({
-    data: data_metagraph_allow_spends,
-  });
-  await prisma.metagraph_spend_transactions.createMany({
-    data: data_metagraph_spend_transactions,
-  });
-  await prisma.metagraph_expired_spend_transactions.createMany({
-    data: data_metagraph_expired_spend_transactions,
-  });
-};
-
-beforeAll(async () => {
-  await seedData();
-});
 
 // Helper validations
 const validateDagAllowSpend = (entry) => {
